Revert "feat(IEX-910): Add method to fetch a drop image object (#90)"
This reverts commit ba6f9ece160c22a08e9bdc84427accd48c923b23.
