+++ examples/drops/backend/src/index.ts
@@ -1,9 +1,8 @@
 import { DropsClient } from '@poap-xyz/drops';
 import { PoapCompass, PoapDropApi } from '@poap-xyz/providers';
-import { create_drop } from './methods/create_drop';
-import { fetch_drop_image } from './methods/fetch_drop_image';
 import { fetch_multiple_drops } from './methods/fetch_multiple_drops';
 import { fetch_single_drop } from './methods/fetch_single_drop';
+import { create_drop } from './methods/create_drop';
 import { getRequiredEnvVar } from './methods/get_required_env_var';
 
 import dotenv from 'dotenv';
@@ -26,8 +25,6 @@ async function main(): Promise<void> {
   await fetch_multiple_drops(client);
   // One Drop by id
   await fetch_single_drop(client);
-  // A Drop Image
-  await fetch_drop_image(client);
   // Create Drop
   await create_drop(client);
 }
