+++ packages/drops/src/queries/index.ts
@@ -1,2 +1,1 @@
 export * from './PaginatedDrop';
-export * from './DropImage';
